+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 * Remove sensitive debug info
+* Avoid calls if project ID not set
 
 ### 1.0.0: June 20th, 2018
 * Creation of the project